refactor: バリデーションのstructを作成。FirebaseClientのメソッドが持っていたバリデーションをValidationへ移動。
